Modify : 아이디 찾기 파일 병합 및 컴포넌트, 함수 수정
