--- src/pages/Findid/FindidPage.scss
@@ -1,30 +1,0 @@
-.findid {
-  margin-top: 100px;
-  padding-bottom: 285px;
-
-  .findidTitle {
-    text-align: center;
-    margin-bottom: 20px;
-  }
-
-  .findidContainer {
-    margin: 0 auto;
-    padding: 20px;
-    width: 450px;
-    border: 1px solid lightgrey;
-
-    .findidMethod {
-      margin-bottom: 40px;
-      text-align: center;
-
-      input {
-        margin-right: 3px;
-        margin-left: 10px;
-      }
-
-      .selectfindWay {
-        margin-right: 5px;
-      }
-    }
-  }
-}
